+++ src/common/utils/deployTips.ts
@@ -21,7 +21,10 @@ export function analyzeDeployErrorLogs(log: string, includeInLog = true): any {
   const updatedLogLines = returnErrorLines(logRes);
   updatedLogLines.forEach((logLine, index) => {
     logResLines.push(logLine);
-    if (logLine.startsWith("Error") && !(updatedLogLines[index + 1] && !updatedLogLines[index + 1].startsWith("Error"))) {
+    if (
+      (logLine.startsWith("Error") || logLine.startsWith(" Error")) &&
+      !(updatedLogLines[index + 1] && !updatedLogLines[index + 1].startsWith("Error"))
+    ) {
       logResLines.push(c.yellow("No sfdx-hardis tip to solve this error. Try google ?"));
       logResLines.push(c.yellow(""));
     }
@@ -90,5 +93,5 @@ function matchesTip(tipDefinition: any, includeInLog = true): boolean | any {
 }
 
 function returnErrorLines(strIn) {
-  return strIn.split(/\r?\n/).filter((str) => str.startsWith("Error") || str.startsWith(firstYellowChar));
+  return strIn.split(/\r?\n/).filter((str) => str.startsWith("Error") || str.startsWith(" Error") || str.startsWith(firstYellowChar));
 }