+++ src/common/utils/deployTipsList.ts
@@ -79,6 +79,15 @@ Example of XML you have to remove in {1}:
 </folderShares>`,
     },
     {
+      name: "can-not-update-field-to-something-else",
+      label: "Cannot update a field to a Summary from something else",
+      expressionRegex: [/Error (.*) Cannot update a field to a (.*) from something else/gm],
+      tip: `You probably updated the type of field {1} to type {2}, and Salesforce does not allows that with deployments. You can:
+- Try to manually change the type of {1} directly in target org, but it may not be technically possible
+- Delete field {1} in target org: it will be recreated after deployment (but you will loose data on existing records, so be careful if your target is a production org)
+- Create another field with desired type and manage data recovery if the target is a production org`,
+    },
+    {
       name: "custom-object-not-found",
       label: "Custom object not found",
       expressionRegex: [/Error (.*) In field: field - no CustomObject named (.*) found/gm],